requirements.txt
django>=1.9.6
factory_boy>=2.7.0
